1.后台添加雷达告警信息 2.对后台用户进行授权 3.添加了用户邮箱信息

+++ src/main/java/com/microthings/monitor_management/mapper/UserMapper.java
@@ -7,7 +7,7 @@ import org.apache.ibatis.annotations.Param;
 import java.util.List;
 
 public interface UserMapper {
-    int countByExample(UserExample example);
+    long countByExample(UserExample example);
 
     int deleteByExample(UserExample example);
